Fixed: app stops receiving inotify events after wakeup from sleep

+++ ElapsedTimer.cpp
@@ -5,15 +5,15 @@ namespace cornus {
 ElapsedTimer::ElapsedTimer() {}
 ElapsedTimer::~ElapsedTimer() {}
 
-inline i8 GetDiffMs(const timespec &end, const timespec &start) {
+inline i64 GetDiffMs(const timespec &end, const timespec &start) {
 	return (end.tv_sec - start.tv_sec) * 1000L + (end.tv_nsec - start.tv_nsec) / 1000000L;
 }
 
-inline i8 GetDiffMc(const timespec &end, const timespec &start) {
+inline i64 GetDiffMc(const timespec &end, const timespec &start) {
 	return (end.tv_sec - start.tv_sec) * 1000000L + (end.tv_nsec - start.tv_nsec) / 1000L;
 }
 
-inline i8 GetDiffNano(const timespec &end, const timespec &start) {
+inline i64 GetDiffNano(const timespec &end, const timespec &start) {
 	return (end.tv_sec - start.tv_sec) * 1000000000L + (end.tv_nsec - start.tv_nsec);
 }
 
@@ -27,7 +27,7 @@ void ElapsedTimer::Continue(const cornus::Reset r)
 		mtl_status(errno);
 }
 
-i8 ElapsedTimer::elapsed_nano()
+i64 ElapsedTimer::elapsed_nano()
 {
 	const bool paused = (start_.tv_sec == 0 && start_.tv_nsec == 0);
 	if (paused)
@@ -38,7 +38,7 @@ i8 ElapsedTimer::elapsed_nano()
 	if (status != 0)
 		mtl_status(errno);
 	
-	const i8 extra = GetDiffNano(now, start_);
+	const i64 extra = GetDiffNano(now, start_);
 	return worked_time_nano_ + extra;
 }
